seperated cards in activity overviews into seperate modules
