--- client/src/app/components/ActivityCard.module.css
@@ -1,12 +1,0 @@
-
-.activityList {
-  margin-top: 1rem;
-
-}
-
-
-
-.listItem{
-    /* color: aqua; */
-    padding: 0.5rem;
-}
