+++ client/src/app/components/ActivityOverviews.module.css
@@ -1,10 +1,11 @@
 
 
 .activityOverviews {
+   
     display: flex;
     flex-direction: column;
     row-gap: 24px;
-    width: 80%;
+    width: 30rem
 }
