+++ client/src/app/components/ClusterCard.module.css
@@ -1,0 +1,18 @@
+.activityList {
+    margin-top: 1rem;
+    max-height: 250px;
+    overflow: auto;
+  
+  }
+  
+
+  .listItem{
+      padding: 0.5rem;
+  }
+  
+  .lastUpdated {
+    font-size: .8rem;
+    color: rgb(193, 193, 193);
+  }
+  
+
